rewrite++

+++ curtana.py
@@ -2,10 +2,7 @@
 # By justaprudev
 
 import json
-import subprocess
 import shutil
-import asyncio
-import requests
 import datetime
 from pathlib import Path
 from telethon.extensions import html
@@ -14,13 +11,13 @@ from jinja2 import Environment, FileSystemLoader
 
 NAME = Path(__file__).stem
 DEFAULT_DATA = {
-        "chats": ["@curtanacloud"],
+        "chats": ["@updates"],
         "item_types": ["rom", "kernel", "recovery", "port", "gsi"],
         "blocked_items": [],
         "filters": [NAME],
         "git": None
         }
-DATA = polygon.db.get(NAME) or polygon.db.add(name=NAME, value=DEFAULT_DATA) or DEFAULT_DATA
+DATA = polygon.db.get(NAME) or polygon.db.add(name=NAME, value=DEFAULT_DATA)
 CWD = Path(DATA["git"].split("/")[-1])
 GLITCH_FOLDER = Path(__file__).parent / "glitch"
 DOMAIN = f"{CWD}.glitch.me"
@@ -30,9 +27,12 @@ DATE = datetime.date.today().strftime("%B %d, %Y")
 @polygon.on(pattern=NAME)
 async def curtana(e):
     if e.sender.username == polygon.user.username: await e.delete()
+    if not DATA:
+        polygon.log(f"Initialization required to use {NAME} pack.")
+        return
     polygon.log(
         f"{DATE} -- its update day!" +
-        f"\nUpdates chat(s): {DATA['chats']}" + 
+        f"\nUpdates chat(s): {DATA['chats']}" +
         f"\nEvent chat: @{e.sender.username}"
     )
     if CWD.exists(): shutil.rmtree(CWD)
@@ -47,17 +47,19 @@ async def curtana(e):
             for content_type in titles:
                 content_list = titles[content_type]
                 if f"#{content_type}" in lower_content and lower_title not in map(str.lower, content_list):
-                    content_list.append(title)    
-                    banner = await get_banner(msg, title)
+                    content_list.append(title)
+                    path = CWD / content_type / title
+                    banner = await get_banner(msg, path)
                     html_content = html.unparse(content, msg.entities)
                     write_webpage(
-                        title=title, 
-                        content=parse_content(html_content.replace(f"#{title}", "", 1)), 
+                        path=path,
+                        title=title,
+                        content=parse_content(html_content.replace(f"#{title}", "", 1)),
                         banner=banner
                         )
     polygon.log(json.dumps(titles, sort_keys=True, indent=1))
     write_webpage(
-        title="index.html",
+        path=GLITCH_FOLDER / "index.html",
         roms=remove_duplicates(titles["rom"] + titles["port"] + titles["gsi"]),
         kernels=remove_duplicates(titles["kernel"]),
         recoveries=remove_duplicates(titles["recovery"]),
@@ -67,58 +69,52 @@ async def curtana(e):
     )
     push_to_glitch(glitch_repository)
     polygon.log("Update completed.")
-    polygon.log(f"Waking up {DOMAIN}..")
-    try: r = requests.head(f"http://{DOMAIN}")
-    except requests.exceptions.ConnectionError: polygon.log(f"Couldn't establish a connection with {DOMAIN}")
-    else: polygon.log(f"Done with {r}")
 
 
-async def get_banner(msg, title):
-    media = await polygon.download_media(msg, f"{CWD}/{title}/")
-    if not media: return ""
-    banner_path = f"{CWD}/{title}/banner"
-    video_ext = ".mp4"
-    if media.endswith(video_ext):
-        banner_path += video_ext
-        banner = f"<video style='border-radius: 10px;' height=255 autoplay loop muted playsinline><source src='https://{DOMAIN}/{title}/banner.mp4' type='video/mp4'></video>"
+async def get_banner(msg, path: Path):
+    try: name = Path(msg.file.name)
+    except AttributeError: return ""
+    path /= f"banner.{name.suffix}"
+    parents = list(map(lambda i: i.stem, path.parents))
+    if name.suffix == ".mp4":
+        banner = f"<video style='border-radius: 10px;' height=255 autoplay loop muted playsinline><source src='https://{DOMAIN}/{parents[1]}/{parents[0]}/{path.name}' type='video/mp4'></video>"
     else:
-        banner_path += ".png"
-        banner = f"<img src='https://{DOMAIN}/{title}/banner.png' height='255'>"
-    Path(media).rename(banner_path)
+        path = path.with_suffix(".png")
+        banner = f"<img src='https://{DOMAIN}/{parents[1]}/{parents[0]}/{path.name}' height='255'>"
+    await polygon.download_media(msg, path)
     return banner
 
 def parse_content(content):
-    replacments = {
-        "<strong>": "", "</strong>": "", "<em>": "",  "</em>": "",
-        "\n": "\n<br>"
+    replacements = {
+        "<strong>": "", "</strong>": "",
+        "<em>": "",  "</em>": "",
+        "\n": "\n                 <br>"
         }
     for i in content.split():
         if i.startswith("@"):
-            replacments[i] = f"<a href=https://t.me/{i[1:]})>{i}</a>"
-    for old, new in replacments.items():
+            replacements[i] = f"<a href=https://t.me/{i[1:]}>{i}</a>"
+    for old, new in replacements.items():
         content = content.replace(old, new)
     return content
 
-def write_webpage(title, **variables):
-    path = f"{CWD}/{title}/index.html"
-    to_path = None
-    if title.endswith(".html"):
-        path = GLITCH_FOLDER / title
-        to_path = f"{CWD}/{title}"
-        jinja2_template = str(open(path, "r").read())
-    else:
+def write_webpage(path: Path, title=None, **variables):
+    if title:
+        path /= "index.html"
         variables["title"] = title
-        jinja2_template = str(open(GLITCH_FOLDER / "template.html", "r").read())
+        jinja2_template = open(GLITCH_FOLDER / "template.html", "r").read()
+    else:
+        jinja2_template = open(path, "r").read()
+        path = CWD / title
     template_object = Environment(
         loader=FileSystemLoader(GLITCH_FOLDER)
         ).from_string(jinja2_template)
-    with open(to_path or path, "w") as f:
-        f.write(template_object.render(**variables)) 
+    with open(path, "w") as f:
+        f.write(template_object.render(**variables))
 
-def clone_from_glitch(repo):
-    repo = Repo.clone_from(DATA["git"], CWD)
+def clone_from_glitch(git_url):
+    repo = Repo.clone_from(git_url, CWD)
     try: repo.head.reset("HEAD~1", index=True, working_tree=True)
-    except GitCommandError: pass
+    except GitCommandError: polygon.log("On root commit.")
     for i in Path(GLITCH_FOLDER).glob("*"):
         if i.is_dir(): shutil.copytree(i, f"{CWD}/{i.stem}", dirs_exist_ok=True)
     return repo
@@ -128,9 +124,13 @@ def push_to_glitch(repo: Repo):
     origin = repo.remote()
     index = repo.index
     index.add("*")
-    files = set(map(lambda i: i.split("/")[0], repo.git.diff("HEAD", name_only=True, diff_filter="A").splitlines()))
-    if "index.html" in files: files.remove("index.html")
-    commit = str(index.commit(f"Added {files}", author=actor, committer=actor, parent_commits=None))
+    commit = str(
+        index.commit(
+        # Glitch uses this commit message
+        "Checkpoint 🚀",
+        author=actor,
+        committer=actor
+        ))
     push = origin.push(force=True)[0]
     polygon.log(f"{DOMAIN} deployed successfully!" if commit[:7] in push.summary else f"Error while deploying:\n{push.summary}\n{commit}")
 
@@ -138,6 +138,7 @@ def is_required_content(content):
     for i in DATA["filters"]:
         if f"#{i}" in content:
             return True
+    return False
 
 def get_random_color():
     from random import randint
